Windows Compatibility

+++ src/test/scripts/tests/AStar.hpp
@@ -17,7 +17,7 @@ namespace Test::Tests {
     public:
 
         template <typename Tm, const size_t Kd, typename Ts = uint32_t>
-        static void Run(const std::array<long unsigned, Kd> _dimensions) {
+        static void Run(const std::array<size_t, Kd> _dimensions) {
 
             #define HEURISTIC CHDR::Heuristics<Kd, Ts>::ManhattanDistance
